+++ backend/zane_api/temporal/workflows.py
@@ -8,6 +8,7 @@ from temporalio.common import RetryPolicy
 from temporalio.exceptions import ApplicationError, ActivityError
 
 from .shared import (
+    DeploymentCreateConfigsResult,
     DeploymentHealthcheckResult,
     SimpleDeploymentDetails,
     ArchivedServiceDetails,
@@ -115,6 +116,7 @@ class RemoveProjectResourcesWorkflow:
 class DockerDeploymentStep(Enum):
     INITIALIZED = auto()
     VOLUMES_CREATED = auto()
+    CONFIGS_CREATED = auto()
     PREVIOUS_DEPLOYMENT_SCALED_DOWN = auto()
     SWARM_SERVICE_CREATED = auto()
     DEPLOYMENT_EXPOSED_TO_HTTP = auto()
@@ -247,6 +249,11 @@ class DeployDockerServiceWorkflow:
                     retry_policy=self.retry_policy,
                 )
 
+            if await check_for_cancellation(DockerDeploymentStep.VOLUMES_CREATED):
+                return await self.handle_cancellation(
+                    deployment, DockerDeploymentStep.VOLUMES_CREATED
+                )
+
             if len(service.configs) > 0:
                 self.created_configs = await workflow.execute_activity_method(
                     DockerSwarmActivities.create_docker_configs_for_service,
@@ -255,9 +262,9 @@ class DeployDockerServiceWorkflow:
                     retry_policy=self.retry_policy,
                 )
 
-            if await check_for_cancellation(DockerDeploymentStep.VOLUMES_CREATED):
+            if await check_for_cancellation(DockerDeploymentStep.CONFIGS_CREATED):
                 return await self.handle_cancellation(
-                    deployment, DockerDeploymentStep.VOLUMES_CREATED
+                    deployment, DockerDeploymentStep.CONFIGS_CREATED
                 )
 
             if (
@@ -496,6 +503,21 @@ class DeployDockerServiceWorkflow:
                     retry_policy=self.retry_policy,
                 )
         if (
+            last_completed_step >= DockerDeploymentStep.CONFIGS_CREATED
+            and len(self.created_configs) > 0
+        ):
+            await workflow.execute_activity_method(
+                DockerSwarmActivities.delete_created_configs,
+                DeploymentCreateConfigsResult(
+                    deployment_hash=deployment.hash,
+                    service_id=deployment.service.id,
+                    created_configs=self.created_configs,
+                ),
+                start_to_close_timeout=timedelta(seconds=5),
+                retry_policy=self.retry_policy,
+            )
+
+        if (
             last_completed_step >= DockerDeploymentStep.VOLUMES_CREATED
             and len(self.created_volumes) > 0
         ):
@@ -726,6 +748,7 @@ def get_workflows_and_activities():
             swarm_activities.get_previous_production_deployment,
             swarm_activities.scale_back_service_deployment,
             swarm_activities.create_deployment_healthcheck_schedule,
+            swarm_activities.delete_created_configs,
             monitor_activities.save_deployment_status,
             monitor_activities.run_deployment_monitor_healthcheck,
             cleanup_activites.cleanup_simple_logs,
